Organised SourceLocator folder into relevance subfolders
