+++ src/Reflection/ReflectionFunction.php
@@ -4,8 +4,8 @@ namespace BetterReflection\Reflection;
 
 use BetterReflection\Reflector\FunctionReflector;
 use BetterReflection\Reflector\Reflector;
-use BetterReflection\SourceLocator\AutoloadSourceLocator;
-use BetterReflection\SourceLocator\LocatedSource;
+use BetterReflection\SourceLocator\Type\AutoloadSourceLocator;
+use BetterReflection\SourceLocator\Located\LocatedSource;
 use PhpParser\Node\Stmt\Function_ as FunctionNode;
 use PhpParser\Node\Stmt\Namespace_ as NamespaceNode;
