+++ src/Reflector/Generic.php
@@ -5,10 +5,8 @@ namespace BetterReflection\Reflector;
 use BetterReflection\Identifier\Identifier;
 use BetterReflection\Identifier\IdentifierType;
 use BetterReflection\Reflection\ReflectionFunction;
-use BetterReflection\SourceLocator\AggregateSourceLocator;
-use BetterReflection\SourceLocator\LocatedSource;
-use BetterReflection\SourceLocator\PhpInternalSourceLocator;
-use BetterReflection\SourceLocator\SourceLocator;
+use BetterReflection\SourceLocator\Located\LocatedSource;
+use BetterReflection\SourceLocator\Type\SourceLocator;
 use BetterReflection\Reflection\ReflectionClass;
 use BetterReflection\Reflection\Reflection;
 use PhpParser\Parser;
