+++ src/SourceLocator/Reflection/SourceStubber.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BetterReflection\SourceLocator;
+namespace BetterReflection\SourceLocator\Reflection;
 
 use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\Interface_;
