+++ src/SourceLocator/Type/AggregateSourceLocator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BetterReflection\SourceLocator;
+namespace BetterReflection\SourceLocator\Type;
 
 use BetterReflection\Identifier\Identifier;
