+++ test/unit/Reflection/ReflectionObjectTest.php
@@ -6,8 +6,8 @@ use BetterReflection\Reflection\ReflectionClass;
 use BetterReflection\Reflection\ReflectionObject;
 use BetterReflection\Reflection\ReflectionProperty;
 use BetterReflection\Reflector\Reflector;
-use BetterReflection\SourceLocator\EvaledLocatedSource;
-use BetterReflection\SourceLocator\LocatedSource;
+use BetterReflection\SourceLocator\Located\EvaledLocatedSource;
+use BetterReflection\SourceLocator\Located\LocatedSource;
 use BetterReflectionTest\Fixture\ClassForHinting;
 use PhpParser\Lexer;
 use PhpParser\Node\Stmt\ClassLike;
