+++ test/unit/TypesFinder/FindParameterTypeTest.php
@@ -6,8 +6,8 @@ use BetterReflection\Reflection\ReflectionClass;
 use BetterReflection\Reflection\ReflectionFunction;
 use BetterReflection\Reflection\ReflectionMethod;
 use BetterReflection\Reflector\ClassReflector;
-use BetterReflection\SourceLocator\LocatedSource;
-use BetterReflection\SourceLocator\StringSourceLocator;
+use BetterReflection\SourceLocator\Located\LocatedSource;
+use BetterReflection\SourceLocator\Type\StringSourceLocator;
 use BetterReflection\TypesFinder\FindParameterType;
 use PhpParser\Node\Param as ParamNode;
 use phpDocumentor\Reflection\Types;
